Planning using STRIPS in PROLOG
An agent close to the microwave oven in the kitchen needs to move to the refrigrator, take out a pizza/chicken from the refrigrator, take it back to the microwave, bake it and eat it.
In this program, STRIPS give us the plan for the agent to achieve the goal of being "not_hungry".